validate and upload image

+++ app/Http/Controllers/userController.php
@@ -2,9 +2,11 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\UserRequest;
 use App\Models\Room;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 
 class userController extends Controller
 {
@@ -25,24 +27,40 @@ class userController extends Controller
         }
         return view('admin.user.Form', ['rooms' => $rooms]);
     }
-    public function store(Request $request, User $id = null)
+    public function store(UserRequest $request, User $id = null)
     {
+        $data = $request->except('_token');
+        if ($request->hasFile('avatar')) {
+            //su ly ten file
+            $avatar = $request->avatar;
+            $avatarName = $avatar->hashName();
+            // dd($avatarName);
+            $avatarName = $request->username . '_' . $avatarName;
+            //luu file vaof trong bo nho
+            $data['avatar'] = $avatar->storeAs('images/users', $avatarName);
+        } else {
+            $data['avatar'] = '';
+        }
         if ($id) {
-            $data = $request->except('_token');
+
             $data['role'] = 1;
             $data['status'] = 1;
             $id->update($data);
             return redirect()->route('users.list');
         }
-        $data = $request->except('_token');
         $data['role'] = 1;
         $data['status'] = 1;
+
         User::create($data);
-        return redirect()->back();
+        return redirect()->route('users.list');
     }
     public function delete($id)
     {
-        User::destroy($id);
+        $user = User::find($id);
+        $image_path = $user->avatar;
+        if (Storage::delete($image_path)) {
+            User::destroy($id);
+        }
         return redirect()->back();
     }
 }
